+++ NEWS.md
@@ -1,5 +1,9 @@
 # normality 0.0.2
 
+* Modified output format.
+* Implement `Shapiro_Wilk_test()` (original, Francia, and Royston versions).
+* Implement a handy wrapper: `is_normal()`.
+
 # normality 0.0.1
 
 * Initial CRAN submission.